model/requirements.txt
numpy
Pillow
torch
torchvision
matplotlib
opencv-python
